Populate message_count using MR job
